Flask-Currency-Conversion-Task/requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
peewee==3.14.4
python-dotenv==0.17.1
requests==2.25.1
urllib3==1.26.5
Werkzeug==2.0.1
